[common/dal] refactor: move DataAccessor and ObjectAccessor to standalone crate common/dal so that other crate can use these accessor without depending on databend-query
