+++ common/dal/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "common-dal"
+version = "0.1.0"
+authors = ["Databend Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+common-base = {path = "../base"}
+common-exception = {path = "../exception"}
+
+async-compat = "0.2.1"
+async-trait = "0.1"
+bytes = "1"
+futures = "0.3"
+rusoto_core = "0.47.0"
+rusoto_s3 = "0.47.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+[dev-dependencies]
+pretty_assertions = "1.0"
+rand = "0.8.4"
