+++ query/src/sessions/context.rs
@@ -22,6 +22,10 @@ use common_base::tokio::task::JoinHandle;
 use common_base::ProgressCallback;
 use common_base::ProgressValues;
 use common_base::TrySpawn;
+use common_dal::DataAccessor;
+use common_dal::Local;
+use common_dal::StorageScheme;
+use common_dal::S3;
 use common_exception::ErrorCode;
 use common_exception::Result;
 use common_infallible::RwLock;
@@ -40,10 +44,6 @@ use crate::catalogs::TableFunctionMeta;
 use crate::catalogs::TableMeta;
 use crate::clusters::ClusterRef;
 use crate::configs::Config;
-use crate::datasources::dal::DataAccessor;
-use crate::datasources::dal::Local;
-use crate::datasources::dal::StorageScheme;
-use crate::datasources::dal::S3;
 use crate::datasources::table_func_engine::TableArgs;
 use crate::sessions::context_shared::DatabendQueryContextShared;
 use crate::sessions::SessionManagerRef;
